Move autologin logic to LoginScreen instead of App
